project2/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
python-dotenv==1.0.1
openai==1.14.3
pytz==2024.1
